io files were renamed for the data tokenization template
